chore(checker): split call parameter display helpers
